Rename file name - LoadBalancerLabel
